Manage rawhide images
To download rawhide tarball we need to replace the
version number (ie 34) by the string Rawhide. To
do that we query bodhi to get all the pending releases
and find out which number is the greater (ie rawhide).
Then when building the url to download the tarballs we
check if the release is rawhide or not.

Signed-off-by: Clement Verna <cverna@tutanota.com>